🚚 rename event_bus module to bus in event_store crate
