+++ marketplace-event-store/src/bus.rs
